+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fuse reparameterable layers before save by `@hglee98` in [PR 531](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/531)
 - Add ONNX model evaluation feature by `@illian01` in [PR 539](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/539)
 - Add ONNX export tool by `@illian01` in [PR 541](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/541)
+- Add VOC 2012 dataset auto-downloader by `@illian01` in [PR 546](https://github.com/Nota-NetsPresso/netspresso-trainer/pull/546)
 
 ## Bug Fixes: